# [ignore] check symints in dispatcher