--- a/aten/src/ATen/NestedTensorImpl.h
+++ b/aten/src/ATen/NestedTensorImpl.h
@@ -169,7 +169,7 @@ struct TORCH_API NestedTensorImpl : public c10::TensorImpl {
   //       we can compute it from `nested_sizes_`
   mutable c10::optional<std::vector<int64_t>> opt_sizes_;
 
-  mutable std::vector<c10::SymInt> size_with_singleton_;
+  mutable c10::optional<std::vector<c10::SymInt>> sym_sizes_;
 
   template <typename VariableVersion>
   c10::intrusive_ptr<TensorImpl> shallow_copy_and_detach_core(