--- a/aten/src/ATen/core/SingletonSymNodeImpl.h
+++ b/aten/src/ATen/core/SingletonSymNodeImpl.h
@@ -6,6 +6,7 @@
 #include <c10/util/Exception.h>
 #include <c10/util/Optional.h>
 #include <c10/util/intrusive_ptr.h>
+#include <ATen/core/TensorBody.h>
 #include <cstdint>
 #include <string>
 
@@ -34,12 +35,36 @@ namespace c10 {
 // During tracing the strides of the outputs need to be a function of the size
 // and strides of the inputs so it is important that SingletonSymNode itself is
 // able to express this.
+enum class SingletonType { PYTHON, CPP };
+
 class TORCH_API SingletonSymNodeImpl : public SymNodeImpl {
  public:
   // CAUTION: you should probably not be constructing these directly; please
-  // the higher-level API in python instead (TODO: actually introduce that).
-  explicit SingletonSymNodeImpl(int64_t val, int64_t coeff)
-      : val_(val), coeff_(coeff) {}
+  // the higher-level API in python instead.
+  explicit SingletonSymNodeImpl(
+      int64_t val,
+      int64_t coeff,
+      at::Tensor vec,
+      int64_t sum_vec,
+      SingletonType type)
+      : val_(val), coeff_(coeff), vec_(std::move(vec)), sum_vec_(sum_vec), type_(type) {
+    if (type == SingletonType::PYTHON) {
+      key_set_ =
+          c10::DispatchKeySet({c10::DispatchKey::Python, c10::DispatchKey::PythonTLSSnapshot});
+    } else if (type == SingletonType::CPP) {
+      // NB: Currently, if SingletonType::CPP is passed, that means that the
+      //     singleton is only meant to be used to ferry nested_tensor_size
+      //     metadata from forward to use in backward. In this case val and
+      //     should be set to some dummy values, and we should make sure to
+      //     error if such an instance is used for anything that relies on val
+      //     and coeff being valid.
+      TORCH_INTERNAL_ASSERT(val == 0 && coeff == 1 && sum_vec == -1);
+      // NB: Since we possibly don't have python instead of relying on torch
+      //     dispatch, we dispatch to the NestedTensor kernel directly.
+      // NB: we can potentially add the AutogradNestedTensor key
+      key_set_ = c10::DispatchKeySet({c10::DispatchKey::NestedTensor});
+    }
+  }
 
   bool bool_() override {
     return false;
@@ -55,6 +80,10 @@ class TORCH_API SingletonSymNodeImpl : public SymNodeImpl {
 
   bool is_bool() override {
     return false;
+  }
+
+  bool is_singleton() override {
+    return true;
   }
 
   bool has_hint() override {
@@ -131,15 +160,30 @@ class TORCH_API SingletonSymNodeImpl : public SymNodeImpl {
   c10::SymNode mul(const c10::SymNode& other) override;
 
   c10::optional<int64_t> singleton_int() override {
+    TORCH_INTERNAL_ASSERT(type_ == SingletonType::PYTHON);
     return val_;
   }
 
   c10::optional<int64_t> singleton_coeff() override {
+    TORCH_INTERNAL_ASSERT(type_ == SingletonType::PYTHON);
     return coeff_;
+  }
+
+  c10::TensorImpl* singleton_vec() override {
+    return vec_.unsafeGetTensorImpl();
+  }
+
+  int64_t singleton_sum_vec() override {
+    TORCH_INTERNAL_ASSERT(type_ == SingletonType::PYTHON);
+    return sum_vec_;
   }
 
   bool is_symbolic() override {
     return false;
+  }
+
+  c10::DispatchKeySet key_set() override {
+    return key_set_;
   }
 
 #define DEFINE_BINARY_NOT_SUPPORTED(name)                           \
@@ -177,6 +221,15 @@ class TORCH_API SingletonSymNodeImpl : public SymNodeImpl {
  private:
   int64_t val_;
   int64_t coeff_;
+  at::Tensor vec_;
+  int64_t sum_vec_;
+  SingletonType type_;
+  c10::DispatchKeySet key_set_;
 };
+
+// If we would like to have singleton_vec() as a virtual method, it must return
+// a TensorImpl* instead of a normal Tensor. We then rely on this helper to
+// wrap that back into a Tensor.
+TORCH_API at::Tensor get_singleton_vec(const c10::SymNode& node);
 
 } // namespace c10