--- a/aten/src/ATen/native/nested/NestedTensorUtils.h
+++ b/aten/src/ATen/native/nested/NestedTensorUtils.h
@@ -411,5 +411,7 @@ inline at::Tensor map_nested_tensor(F&& fn, A... a) {
       c10::nullopt);
 }
 
+Tensor get_nested_sizes_from_sym_sizes(SymIntArrayRef size);
+
 } // namespace native
 } // namespace at