… aten/src/ATen/native/native_functions.yaml
… test/cpp/api/singleton_int.cpp
… test/dynamo/test_subclasses.py
… test/test_dynamic_shapes.py
… test/test_nestedtensor.py
… tools/autograd/derivatives.yaml
… torch/_C/__init__.pyi.in
… torch/fx/experimental/sym_node.py
… torch/fx/experimental/symbolic_shapes.py
… torch/nested/_internal/nested_tensor.py
… torch/nested/_internal/ops.py
… torchgen/gen.py